+++ src/memory_manager.rs
